Switch resolving order if source not specified in `load_adapter()` (#681)
If `source` is not specified in `load_adapter()`, will switch priorities
to first search on HF Hub and only on AH if not found there. Related to
AH deprecation (#668).

This also fixes an bug with auto-resolving adapters for which no model
index file is given in AH (e.g. currently for Llama-2).