Add Support for Gradient Checkpointing (#759)
# Add Support for Gradient Checkpointing
This PR adds support for gradient checkpointing.

Gradient checkpointing is a technique that trades computation for memory by recomputing intermediate activations during the backward pass instead of storing them. This is particularly useful when training large models. Because we
recompute values during the backpropagation, we need to preserve the original ForwardContext in this phase. I solved this by overwriting the `gradient_checkpointing_enable` function so that the checkpoint function
receives the current ForwardContext as the backward pass context manager.

---------

Co-authored-by: calpt <calpt@mail.de>