Option for initializing adapters with identical weights  (#786)
Adresses #653 and introduces an additional parameter in the adapter
config that allows to fix a seed during adapter weights initialization
**for every layer**.

This means that the seed is reset for every layer to the specified
value, which leads to all adapter modules having the same weights upon
initialization.

The PR adds this option as an additonal argument `init_weights_seed` in
the adapter config and also provides an additional test that checks for
identical weight initialization between multiple adapters and multiple
models.

Closes #653