Adapter Interface: Validate Values (#797)
This PR introduces a validation for the values specified in the adapter
interfaces. We now check that the layers specified actually exist and if
not, we provide a detailed error to the user so that they know what went
wrong.

Before this PR, when users specify layers in the adapter interfaces that
don't exist in the model, there was no error thrown. This left users
unaware that their specified values were invalid.