Implements `Vera` (#763)
This PR implements Vera, a LoRA variant that introduces trainable parameters
`d` and `b` while keeping the LoRA matrices `A` and `B` frozen, random, and
shared across layers.

I've opted to put the Vera implementation under the Lora implementation
(like IA3)

Paper: https://arxiv.org/pdf/2310.11454

This PR includes:

- A new `Vera` implementation under the lora methods file
- A new adapter config named `VeraConfig`
- A shared parameter initialisation function that takes in the
`init_weights` argument set from the `VeraConfig`
- A new criterion inside the `add_adapter` function inside the
`LoRALayer` to check if we should use the `Vera` class. Currently, the
criteria will check to see if the parameters from the `VeraConfig` `d`
and `b` are of type `float`, it's the simplest criteria I could think
of at the moment.
- A new parameter was added inside `add_adapter` inside the
`LoRALayer` as suggested. It will now pass the name of the adapter
inside the `__init__` function.
- A fix for the `average_adapter` for prompt tuning.

---------

Co-authored-by: Leon Engländer <leon.englaender@gmail.com>