Fix invalid .json for adapter notebooks (#812)
While scanning through the notebooks, @julian-fong encountered the
following error:

![image](https://github.com/user-attachments/assets/e6f4bbba-a4b5-4091-b0ff-23d960acbea9)

Given the instructions in the error message on how to fix the issue I
took option two and removed the metadata.widgets key with its content
from the affected notebooks. As a result the notebooks can be rendered
again normally.